refactor: rename component
